backend/requirements.txt
fastapi
uvicorn[standard]
gunicorn
pydantic
pydantic-settings
sqlalchemy
asyncpg
alembic
geoalchemy2
psycopg2-binary
redis
celery
torch
torch-geometric
filterpy
shap
onnxruntime
httpx
rasterio
africastalking
PyJWT
python-multipart
slowapi
python-jose[cryptography]
passlib[bcrypt]
structlog
python-dotenv
pytest
pytest-asyncio
